Refactor ascii functions
Move `encodeAscii` and `decodeAscii` to a file.
